% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_fuze.R
\name{bom_fuze}
\alias{bom_fuze}
\title{Fuze numeric sound wave tibble thumbrints into analysis data model (ADaM)}
\usage{
bom_fuze()
}
\value{
- analysis data set/model (ADaM)
}
\description{
Fuze numeric sound wave tibble thumbrints into analysis data model (ADaM)
}
\examples{
- bom_fuze()
}
