% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_svm_test.R
\name{bom_svm_test}
\alias{bom_svm_test}
\title{Use ADaM-trained SVM to ask a question about any wave's numeric thumbprint}
\usage{
bom_svm_test(COCO)
}
\arguments{
\item{COCO}{- name of the thumbprint file (must have a space after it)}
}
\value{
- color graph where cream = man and red = bird, so you will know the decision
}
\description{
Use ADaM-trained SVM to ask a question about any wave's numeric thumbprint
}
\examples{
- bom_svm_test(COCO = "loon ")
}
