% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_svm_train.R
\name{bom_svm_train}
\alias{bom_svm_train}
\title{Training Support Vector Machine on analysis data model (ADaM)}
\usage{
bom_svm_train()
}
\value{
- graphical and accuracy output that the training was successful
}
\description{
Training Support Vector Machine on analysis data model (ADaM)
}
\examples{
- bom_svm_train()
}
