% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_thmb.R
\name{bom_thmb}
\alias{bom_thmb}
\title{Draw a spectrogram and create a numeric thumbprint
of a sound wave}
\usage{
bom_thmb(ANGEL, COCO, ID)
}
\arguments{
\item{ANGEL}{- "bird" or "man" (who made the sound)}

\item{COCO}{- file name (e.g., "actor1")}

\item{ID}{- 1 to 10 actors, 1 to 10 birds, Eva is 11th man}
}
\value{
- spectrogram, spectrogram jpeg, tibble representing specific sound wave
}
\description{
Draw a spectrogram and create a numeric thumbprint
of a sound wave
}
\examples{
- bom_thmb(ANGEL = "man", COCO = "Eva", ID = 11)
}
