% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_viz.R
\name{bom_viz}
\alias{bom_viz}
\title{Visualizes clusters of birds and men after wave processing
in any amplitude plane}
\usage{
bom_viz(A1, A2)
}
\arguments{
\item{A1}{- Amplitude variable 1 (Amp1 - Am10)}

\item{A2}{- Amplitude variable 2 (Amp1 - Am10)}
}
\value{
- side-by-side plots with and without id labels showing birds and men
}
\description{
Visualizes clusters of birds and men after wave processing
in any amplitude plane
}
\examples{
- bom_viz(A1 = "Amp1", A2 = "Amp2")
}
